refactor: directory name from caml case to underscore
