+++ lib/router.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
-import 'package:messless/screens/technicalLog/qr_scanner_screen.dart';
-import 'package:messless/screens/technicalLog/add_technical_log_entry.dart';
+import 'package:messless/screens/technical_log/qr_scanner_screen.dart';
+import 'package:messless/screens/technical_log/add_technical_log_entry.dart';
 import 'package:messless/screens/equipment/add_equipment.dart';
 import 'package:messless/screens/equipment/equipment.dart';
 import 'package:messless/screens/equipment/equipment_details.dart';
