+++ lib/ws/schema/equipment_storage/equipment_storage.dart
