+++ lib/ws/schema/technical_log_entry/technical_log_entry.dart
